Refactor Section data to be in its own file
